Migrated all results based actions to results controller
