+++ src/Syringe.Service/packages.config
@@ -3,6 +3,8 @@
   <package id="Common.Logging" version="3.0.0" targetFramework="net451" />
   <package id="Common.Logging.Core" version="3.0.0" targetFramework="net451" />
   <package id="Microsoft.AspNet.Cors" version="5.0.0" targetFramework="net451" />
+  <package id="Microsoft.AspNet.Mvc" version="5.2.3" targetFramework="net451" />
+  <package id="Microsoft.AspNet.Razor" version="3.2.3" targetFramework="net451" />
   <package id="Microsoft.AspNet.SignalR.Core" version="2.2.0" targetFramework="net452" />
   <package id="Microsoft.AspNet.SignalR.SelfHost" version="2.2.0" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net451" />
@@ -10,6 +12,7 @@
   <package id="Microsoft.AspNet.WebApi.Owin" version="5.2.3" targetFramework="net451" />
   <package id="Microsoft.AspNet.WebApi.OwinSelfHost" version="5.2.3" targetFramework="net451" />
   <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.3" targetFramework="net451" />
+  <package id="Microsoft.AspNet.WebPages" version="3.2.3" targetFramework="net451" />
   <package id="Microsoft.Owin" version="3.0.1" targetFramework="net451" />
   <package id="Microsoft.Owin.Cors" version="3.0.1" targetFramework="net451" />
   <package id="Microsoft.Owin.Diagnostics" version="3.0.1" targetFramework="net451" />
