+++ src/Syringe.Tests/Unit/Web/ResultsControllerTests.cs
@@ -16,19 +16,27 @@ namespace Syringe.Tests.Unit.Web
     [TestFixture]
     public class ResultsControllerTests
     {
+        private int _id;
         private Mock<ITasksService> _tasksServiceMock;
         private Mock<IUrlHelper> _urlHelperMock;
+        private Mock<ITestService> _testsClient;
+
         private ResultsController _resultsController;
-        private int _id;
 
         [SetUp]
         public void Setup()
         {
+            _id = 0;
             _tasksServiceMock = new Mock<ITasksService>();
             _urlHelperMock = new Mock<IUrlHelper>();
-            _id = 0;
+
+            _testsClient = new Mock<ITestService>();
+            _testsClient.Setup(x => x.GetResultById(It.IsAny<Guid>())).Returns(new TestFileResult());
+            _testsClient.Setup(x => x.GetSummaries()).Returns(new List<TestFileResultSummary>());
+            _testsClient.Setup(x => x.GetSummariesForToday()).Returns(new List<TestFileResultSummary>());
+
             _tasksServiceMock.Setup(x => x.GetRunningTaskDetails(It.IsAny<int>())).Returns((new TaskDetails { Results = new List<TestResult> { new TestResult { ActualUrl = "syringe.com", Position = _id } } }));
-            _resultsController = new ResultsController(_tasksServiceMock.Object, _urlHelperMock.Object);
+            _resultsController = new ResultsController(_tasksServiceMock.Object, _urlHelperMock.Object, _testsClient.Object);
         }
 
         [Test]
@@ -79,6 +87,56 @@ namespace Syringe.Tests.Unit.Web
             // then
             Assert.IsInstanceOf<ResultsViewModel>(actionResult.Model);
             _tasksServiceMock.Verify(x => x.GetRunningTaskDetails(It.IsAny<int>()), Times.Once);
+        }
+
+
+        [Test]
+        public void AllResults_should_return_correct_view_and_model()
+        {
+            // given + when
+            var viewResult = _resultsController.AllResults() as ViewResult;
+
+            // then
+            _testsClient.Verify(x => x.GetSummaries(), Times.Once);
+            Assert.AreEqual("AllResults", viewResult.ViewName);
+            Assert.IsInstanceOf<IEnumerable<TestFileResultSummary>>(viewResult.Model);
+        }
+
+        [Test]
+        public void TodaysResults_should_return_correct_view_and_model()
+        {
+            // given + when
+            var viewResult = _resultsController.TodaysResults() as ViewResult;
+
+            // then
+            _testsClient.Verify(x => x.GetSummariesForToday(), Times.Once);
+            Assert.AreEqual("AllResults", viewResult.ViewName);
+            Assert.IsInstanceOf<IEnumerable<TestFileResultSummary>>(viewResult.Model);
+        }
+
+
+        [Test]
+        public void ViewResult_should_return_correct_view_and_model()
+        {
+            // given + when
+            var viewResult = _resultsController.ViewResult(It.IsAny<Guid>()) as ViewResult;
+
+            // then
+            _testsClient.Verify(x => x.GetResultById(It.IsAny<Guid>()), Times.Once);
+            Assert.AreEqual("ViewResult", viewResult.ViewName);
+            Assert.IsInstanceOf<TestFileResult>(viewResult.Model);
+        }
+
+        [Test]
+        public async void DeleteResult_should_call_delete_methods_and_redirect_to_correct_action()
+        {
+            // given + when
+            var redirectToRouteResult = await _resultsController.DeleteResult(It.IsAny<Guid>()) as RedirectToRouteResult;
+
+            // then
+            _testsClient.Verify(x => x.GetResultById(It.IsAny<Guid>()), Times.Once);
+            _testsClient.Verify(x => x.DeleteResultAsync(It.IsAny<Guid>()), Times.Once);
+            Assert.AreEqual("AllResults", redirectToRouteResult.RouteValues["action"]);
         }
     }
 }
