+++ src/Syringe.Web/Controllers/HomeController.cs
@@ -1,14 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 using System.Web.Mvc;
 using Syringe.Core.Configuration;
 using Syringe.Core.Extensions;
-using Syringe.Core.Helpers;
 using Syringe.Core.Security;
 using Syringe.Core.Services;
-using Syringe.Core.Tests.Results;
 using Syringe.Web.Models;
 
 namespace Syringe.Web.Controllers
@@ -17,25 +14,19 @@ namespace Syringe.Web.Controllers
     public class HomeController : Controller
     {
         private readonly ITestService _testsClient;
-        private readonly IUserContext _userContext;
         private readonly Func<IRunViewModel> _runViewModelFactory;
 		private readonly IHealthCheck _healthCheck;
-	    private readonly IUrlHelper _urlHelper;
 	    private readonly IEnvironmentsService _environmentsService;
 
 	    public HomeController(
             ITestService testsClient,
-            IUserContext userContext,
             Func<IRunViewModel> runViewModelFactory,
-			IHealthCheck healthCheck, 
-            IUrlHelper urlHelper,
+			IHealthCheck healthCheck,
             IEnvironmentsService environmentsService)
         {
             _testsClient = testsClient;
-            _userContext = userContext;
             _runViewModelFactory = runViewModelFactory;
 			_healthCheck = healthCheck;
-	        _urlHelper = urlHelper;
 	        _environmentsService = environmentsService;
         }
 
@@ -83,65 +74,5 @@ namespace Syringe.Web.Controllers
 			_healthCheck.CheckServiceConfiguration();
 			_healthCheck.CheckServiceSwaggerIsRunning();
         }
-
-        public ActionResult AllResults()
-        {
-            return View("AllResults", _testsClient.GetSummaries());
-        }
-
-        public ActionResult TodaysResults()
-        {
-            return View("AllResults", _testsClient.GetSummariesForToday());
-        }
-
-        public ActionResult ViewResult(Guid id)
-        {
-            return View("ViewResult", _testsClient.GetResultById(id));
-        }
-
-		[HttpPost]
-        public async Task<ActionResult> DeleteResult(Guid id)
-        {
-            TestFileResult session = _testsClient.GetResultById(id);
-            await _testsClient.DeleteResultAsync(session.Id);
-
-            return RedirectToAction("AllResults");
-        }
-
-		public ActionResult ViewHtml(Guid testFileResultId, int resultId)
-		{
-			TestFileResult testFileResult = _testsClient.GetResultById(testFileResultId);
-		    TestResult result = testFileResult.TestResults.ElementAtOrDefault(resultId);
-			if (result != null)
-			{
-				string html = result.HttpContent;
-				string baseUrl = _urlHelper.GetBaseUrl(result.ActualUrl);
-				html = _urlHelper.AddUrlBase(baseUrl, html);
-
-				return Content(html);
-			}
-
-			return Content("Result Id not found");
-		}
-
-		public ActionResult ViewHttpLog(Guid testFileResultId, int resultId)
-		{
-			TestFileResult testFileResult = _testsClient.GetResultById(testFileResultId);
-			TestResult result = testFileResult.TestResults.ElementAtOrDefault(resultId);
-            if (result != null)
-				return Content(result.HttpLog, "text/plain");
-
-			return Content("Result Id not found");
-		}
-
-		public ActionResult ViewLog(Guid testFileResultId, int resultId)
-		{
-			TestFileResult testFileResult = _testsClient.GetResultById(testFileResultId);
-			TestResult result = testFileResult.TestResults.ElementAtOrDefault(resultId);
-            if (result != null)
-				return Content(result.Log, "text/plain");
-
-			return Content("Result Id not found");
-		}
 	}
 }
