+++ src/Syringe.Web/Controllers/TestController.cs
@@ -1,10 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Web.Mvc;
-using Microsoft.Ajax.Utilities;
 using Syringe.Core.Extensions;
-using Syringe.Core.Security;
 using Syringe.Core.Services;
 using Syringe.Core.Tests;
 using Syringe.Web.Mappers;
@@ -17,18 +14,15 @@ namespace Syringe.Web.Controllers
 	public class TestController : Controller
 	{
 		private readonly ITestService _testsClient;
-		private readonly IUserContext _userContext;
 		private readonly ITestFileMapper _testFileMapper;
 	    private readonly IEnvironmentsService _environmentsService;
 
 		public TestController(
 			ITestService testsClient,
-			IUserContext userContext,
 			ITestFileMapper testFileMapper,
             IEnvironmentsService environmentsService)
 		{
 			_testsClient = testsClient;
-			_userContext = userContext;
 			_testFileMapper = testFileMapper;
 	        _environmentsService = environmentsService;
 		}
