+++ src/Syringe.Web/Views/Home/Run.cshtml
@@ -21,7 +21,7 @@
 			<a class="btn btn-primary pull-right" target="_blank" href="@Url.Action("View", "Test", new {filename = Model.FileName})">View Tests</a>
 			@using (Html.BeginForm("Run", "Home"))
 			{
-                @Html.HiddenFor(x => x.Environment)
+				@Html.HiddenFor(x => x.Environment)
 				<button title="Run" class="btn btn-primary pull-right" type="submit" name="filename" value="@Model.FileName"> Run again </button>
 			}
 		</div>
