--- src/Syringe.Web/Views/Home/ViewResult.cshtml
@@ -1,116 +1,0 @@
-﻿@model TestFileResult
-@{
-	ViewBag.Title = "Run";
-}
-<div class="container" id="test-result">
-	<div class="row">
-		<div class="col-xs-12">
-			<form action="@Url.Action("DeleteResult", "Home")" method="POST">
-				<input type="hidden" name="id" value="@Model.Id" />
-				<button class="btn btn-danger btn-sm pull-right delete-button">
-					Delete
-				</button>
-			</form>
-			<h1>
-				Results for @Model.Filename
-			</h1>
-			<h4>
-				@Model.StartTime.ToString("ddd dd MMM yyyy HH:mm")
-			</h4>
-
-			<div class="row">
-				<div class="col-md-12">
-					<a href="#" class="btn btn-primary" id="all-tests-count">
-						All Tests
-					</a>
-					<a href="#" class="btn btn-success" id="tests-passed-count">
-						@Model.TotalTestsPassed tests passed
-					</a>
-					<a href="#" class="btn btn-danger" id="tests-failed-count">
-						@Model.TotalTestsFailed tests failed
-					</a>
-                    <div class="btn btn-info" style="cursor: default">
-                        Ran for @Model.TotalRunTime.MinutesAndSecondsFormat()
-                    </div>
-                    <div class="btn btn-material-indigo" style="cursor: default" title="Environment: @Html.AttributeEncode(Model.Environment)">
-                        @(Model.Environment ?? "## MISSING ##")
-                    </div>
-				</div>
-			</div>
-
-			<div class="row">
-				<div class="col-xs-12">
-					<div class="panel-group" id="accordion" role="tablist" aria-multiselectable="true">
-						@foreach (TestResult result in Model.TestResults)
-						{
-							<div class="panel @(result.Success ? "test-passed panel-success" : "test-failed panel-danger")">
-								<div class="panel-heading" role="tab" id="heading@(result.Position)">
-									<h3 class="panel-title">
-										<a role="button" data-toggle="collapse" href="#collapse@(result.Position)" aria-expanded="false" aria-controls="collapse@(result.Position)">
-											@result.Test.Description
-										</a>
-									</h3>
-								</div>
-								<div id="collapse@(result.Position)" class="panel-collapse collapse" role="tabpanel" aria-labelledby="heading@(result.Position)">
-									<div class="panel-body">
-										<ul>
-											<li><a href="@result.ActualUrl" target="_new">@result.ActualUrl</a></li>
-											<li>Took @result.ResponseTime.MinutesAndSecondsFormat()</li>
-											<li>Response Code Success: @result.ResponseCodeSuccess</li>
-										</ul>
-
-										<p>
-											<a class="btn btn-primary" target="_blank" href="@Url.Action("ViewLog", "Home", new {testFileResultId = Model.Id, resultId = result.Position})">View Log</a>
-											<a class="btn btn-primary" target="_blank" href="@Url.Action("ViewHttpLog", "Home", new {testFileResultId = Model.Id, resultId = result.Position})">View HTTP Log</a>
-										    <a class="btn btn-primary" target="_blank" href="@Url.Action("ViewHtml", "Home", new {testFileResultId = Model.Id, resultId = result.Position})">View HTML</a>
-                                            <a class="btn btn-primary" target="_blank" href="@Url.Action("Edit", "Test", new {filename = Model.Filename, position = result.Position})">Edit</a>
-										</p>
-
-										@if (!string.IsNullOrEmpty(result.ExceptionMessage))
-										{
-											<p>
-												<strong>Exception in the test:</strong><br />
-												@result.ExceptionMessage
-											</p>
-										}
-
-										@if (result.ResponseCodeSuccess)
-										{
-
-											<table class="table table-bordered table-striped">
-												<tr>
-													<td>Description</td>
-													<td>Value</td>
-													<td>Assertion type</td>
-													<td>Assertion method</td>
-												</tr>
-
-												@foreach (var item in result.AssertionResults)
-												{
-													<tr class="@(item.Success ? "success" : "danger")">
-														<td>
-															@item.Description
-														</td>
-														<td>
-															@item.Value
-														</td>
-														<td>
-															@item.AssertionType.ToString()
-														</td>
-														<td>
-															@item.AssertionMethod.ToString()
-														</td>
-													</tr>
-												}
-											</table>
-										}
-									</div>
-								</div>
-							</div>
-						}
-					</div>
-				</div>
-			</div>
-		</div>
-	</div>
-</div>
