+++ src/Syringe.Web/Views/Results/AllResults.cshtml
@@ -1,0 +1,45 @@
+﻿@using Syringe.Core.Tests.Results
+@model IEnumerable<TestFileResultSummary>
+@{
+    ViewBag.Title = "Viewing Results";
+}
+<div class="container">
+    <div class="row">
+        <div class="col-xs-12">
+            <h2>All results</h2>
+            <table class="table table-bordered table-striped">
+                <thead>
+                    <tr>
+                        <th>Date</th>
+                        <td>Name</td>
+                        <td>Environment</td>
+                        <td>Result</td>
+                        <td></td>
+                    </tr>
+                </thead>
+                <tbody>
+                    @foreach (TestFileResultSummary session in Model)
+                    {
+                        <tr>
+                            <td>@session.DateRun.ToLocalTime()</td>
+                            <td>
+                                @session.FileName
+                            </td>
+                            <td>
+                                <span class="label label-material-indigo">@session.Environment</span>
+                            </td>
+                            <td>
+                                <span class="label label-success results-label">Total Passed: @session.TotalPassed</span>
+                                <span class="label label-danger results-label">Total Failed: @session.TotalFailed</span>
+                                <span class="label label-info results-label">Run Time: @session.TotalRunTime.MinutesAndSecondsFormat()</span>
+                            </td>
+                            <td>
+                                <a class="btn btn-sm btn-primary" href="@Url.Action("ViewResult", "Results", new {id = session.Id})">View</a>
+                            </td>
+                        </tr>
+                    }
+                </tbody>
+            </table>
+        </div>
+    </div>
+</div>
